requirements.txt
streamlit
google-generativeai
youtube-search
duckduckgo-search
PyPDF2
python-pptx
Pillow